feat: add support for sending Telegram notifications
Adds support for sending notifications via Telegram.

This change introduces libcurl as a dependency to
enable sending HTTP requests to the Telegram Bot API.
The removal of notify-send auto-installation simplifies
the build process and relies on the user to have the
necessary dependencies.

telegram_bot_token variable will need to be configured.